CA-93914: No longer setting params in sm_config on lvoiscsi SR create
For some reason (unknown) the CIM layer was setting a number of sm_config
params, presumably to get around historical issues. The SM layer however
expects these kvps to only be used by itself, and so can get confused.

The fact that the CIM layer was setting the devserial, meant that when SM
went to check if the device existed (on the basis of checking devserial's)
it determined that someone had already added the disk (because it checks
itself too).

Despite this perhaps being unwise, the CIM layer should not touch sm_config,
and given we are not sure why it ever needed too - I'm removing this code.

(Note that due to a bug in the bindings, we have be ingoring passing this
param on for a considerable time anyway).

Signed-off-by: Rob Dobson <rob@rdobson.co.uk>