CA-118036: Revert allocations of string_string maps to be the empty list.
There seems to have been a reason why allocation of empty sets for parameters
like other_config and VCPUs_params were allocated with a single item (as opposed
to no zero items).

With a recent update to the XenAPI xml-rpc parsing library, it has been discovered
that this non-zero initialisation has lead to an empty member being defined
as part of the dictionary:

<member>
    <name>VCPUs_params</name>
    <value>
        <struct>
            <member>
                <name/>
                <value><string/></value>
            </member>
        </struct>
    <value>
</member>

The (new) xml-parsing library (understandably) does not know how to handle
this unamed member, and raises an exception.

This patch fixes the instances where we call an initialisation on a
string_string map to make sure we don't pre-allocate a map with an empty
value.

Signed-off-by: Rob Dobson <rob@rdobson.co.uk>