CA-135188: Cache the Host CN name.
In order to improve performance, use a other_config key to cache the result of obtaining a host CN from a XAPI plugin call. This is OK as the host CN should not change once the host has been installed (embedded in the XAPI SSL cert).

Also update the cert management script to clear the CN from other_config when it is updated by the user.